Oops. Fixed regression introduced by [4594] (change from mutable set to mutable dict for notification-names uniquification): We stopped adding a GRDENotification to the dict (formerly set) when committing edits to a GRDENotification for a first time.
Steps to Reproduce:
1. Click Add.
2. Enter a name.
3. Click Add.

At this point, the GRDENotification created in step 1 has its edit from step 2 committed to it. This was when we should have added the notification to the dictionary, but we didn't. We had previously been adding it to the set, but I commented it out for not knowing what the correct object to add was (before looking up and seeing the forObject parameter to this method, the observeValueForKeyPath:ofObject:change:context: method).