Fixed the warning on launch about the the Growl UDP pathway being unable to bind the socket. This happened since [4803] when Growl was set to listen for forwarded notifications. It didn't happen before that revision because we terminated before the pathways were instantiated (also before any documents were opened, which was the problem it fixed).
More specifically:

Moved -[GrowlApplicationController initSingleton]'s instantiation of the GrowlPathwayController above its call to -preferencesChanged:, because -preferencesChanged: was instantiating the GPC anyway.

In that same section, made -initSingleton responsible for checking whether another GHA was already running.

Made -preferencesChanged: not instantiate the GPC when the quitAfterOpen bit is set.

Changed GrowlApplicationBridgePathway to not set the quitAfterOpen attribute itself. That's GAC's job, now.