Restore the use of mach_error_string to obtain a description of the error code returned by mach_inject.
This previously wasn't working under x86_64; that's because we weren't including the header that contains mach_error_string's declaration, so the compiler was treating its return value as an int. That's really bad when it actually returns a pointer and a pointer is bigger than an int. Including the header means the compiler now handles the return value properly, fixing the crash that had ensued.