Use FSEvents to detect when the user installs plug-ins in the plug-ins folder.
Currently has no visible effect in the prefpane, but does mean that you can drop a display plug-in into the Plugins folder, then open the prefpane, and be able to preview the display. Previously, the preview wouldn't work until you restarted Growl (or moved the plug-in back out of the folder and then double-clicked it).

We also detect when the user removes a plug-in from the folder, although we currently have no use for this knowledge.