Workaround for the 1-pt-tall-notification-window problem (#108).
The problem seems to be that WebKit hasn't finished laying out the content, even by the time of webView:didFinishLoadForFrame:. It may be that WebKit is getting distracted and laying out other notifications before it finishes the first.

An apparent fix is to only attempt to get and use the size after a delay, at least when the size isn't already in yet. I'm not yet certain that this is a real fix, but it at least makes the problem much less frequent, to the point that I'm no longer able to reproduce it.