Add code to prevent users from just hitting the Enter button to skip past the warning message, consisting of a 15-second delay before the OK button is enabled.
There's more manual work here than usual: Searching out the button from inside the view hierarchy rather than using an outlet; and enabling it, both initially and after the timer expires, programmatically. All of this is to make the patcher resistant to users hacking the nib in IB. The idea is for them to need to download the source and compile the whole thing in Xcode in order to bypass the delay.